move getConversationWithoutContent to its own file
